+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/components/export.js
@@ -38,11 +38,19 @@ ResultNumbersView = Backbone.Marionette.ItemView.extend({
 
     onShow: function() {
 
-        var datasource_whitelist = ['ops'];
+        // compute crawler by datasource
+        var crawler_class;
         var datasource = this.model.get('datasource');
-        if (!_(datasource_whitelist).contains(datasource)) {
-            this.user_message('Fetching result publication numbers for datasource "' + datasource + '" not implemented yet.', 'error');
+        if (datasource == 'ops') {
+            crawler_class = OpsPublishedDataCrawler;
+
+        } else if (datasource == 'depatisnet') {
+            crawler_class = DepatisnetCrawler;
+
+        } else {
+            this.user_message('Fetching publication numbers for datasource "' + datasource + '" not implemented yet.', 'error');
             return;
+
         }
 
         var query = this.model.get('query_origin');
@@ -50,32 +58,31 @@ ResultNumbersView = Backbone.Marionette.ItemView.extend({
         this.indicate_activity(true);
         this.user_message('Fetching result numbers for query "' + query + '", please stand by. &nbsp; <i class="spinner icon-refresh icon-spin"></i>', 'info');
 
-        var url = '/api/ops/published-data/crawl/pub-number?query=' + query;
         var _this = this;
-        $.ajax({url: url, async: true})
-            .success(function(payload) {
-                if (payload) {
+        var crawler = new crawler_class({constituents: 'pub-number', query: query})
+        crawler.crawl().then(function(numberlist) {
 
-                    // display numberlist payload in textarea
-                    var numberlist = payload['ops:world-patent-data']['ops:biblio-search']['ops:search-result']['publication-numbers'];
-                    var numberlist_string = numberlist.join('\n');
-                    $('#result-numbers-content').val(numberlist_string);
+                // transfer data
+                var numberlist_string = numberlist.join('\n');
+                $('#result-numbers-content').val(numberlist_string);
 
-                    _this.setup_copy_button(numberlist_string);
-
-                    // notify user
-                    _this.indicate_activity(false);
-                    _this.user_message('Result publication numbers fetched successfully. Hits: ' + numberlist.length, 'success');
-                }
-            }).error(function(error) {
+                _this.setup_copy_button(numberlist_string);
 
                 // notify user
+                _this.indicate_activity(false);
+                _this.user_message('Result publication numbers fetched successfully. Hits: ' + numberlist.length, 'success');
+
+            }).fail(function(message, error) {
+
+                // notify user
+                _this.indicate_activity(false);
+
                 var message =
                     'Error while crawling numberlist for query "' + query +
-                        '" at datasource "' + _this.model.get('datasource') + '".';
-                console.warn(message, error);
-                _this.indicate_activity(false);
+                        '" at datasource "' + _this.model.get('datasource') + '".' + '<br/>' + message;
                 _this.user_message(message, 'error');
+                console.warn(message, error);
+
             });
 
     },