+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - middleware: number normalization: AT362828E should be returned as AT362828T for querying at OPS
 - middleware/ui: ops upstream datasource crawler for fetching complete list of publication numbers, user interface
 - ui: fix “full-cycle” for firefox
+- middleware/ui: numberlist crawler for depatisnet
 
 0.84.0
 ======